[FLINK-27445] Create Flink Standalone Service
