+++ flink-kubernetes-operator/src/main/java/org/apache/flink/kubernetes/operator/crd/spec/FlinkDeploymentSpec.java
@@ -70,4 +70,7 @@ public class FlinkDeploymentSpec extends AbstractFlinkSpec {
      * configContent.
      */
     private Map<String, String> logConfiguration;
+
+    /** Deployment mode of the Flink cluster, native or standalone. */
+    private KubernetesDeploymentMode mode;
 }
